requirements.txt
beautifulsoup4==4.6.0
chardet==3.0.4
cssselect==1.0.1
html2text==2017.10.4
ImageHash==4.0
lxml==3.8.0
numpy==1.14.0
Pillow==5.0.0
PySocks==1.6.7
PyWavelets==0.5.2
scipy==1.0.0
selenium==3.8.1
SerpScrap==0.10.1
six==1.11.0
SQLAlchemy==1.1.15
